+++ src/React/SocketClient/Connector.php
@@ -5,7 +5,7 @@ namespace React\SocketClient;
 use React\EventLoop\LoopInterface;
 use React\Dns\Resolver\Resolver;
 use React\Stream\Stream;
-use React\Promise\When;
+use React\Promise;
 use React\Promise\Deferred;
 
 class Connector implements ConnectorInterface
@@ -35,7 +35,7 @@ class Connector implements ConnectorInterface
         $socket = stream_socket_client($url, $errno, $errstr, 0, STREAM_CLIENT_CONNECT | STREAM_CLIENT_ASYNC_CONNECT);
 
         if (!$socket) {
-            return When::reject(new \RuntimeException(
+            return Promise\reject(new \RuntimeException(
                 sprintf("connection to %s:%d failed: %s", $address, $port, $errstr),
                 $errno
             ));
@@ -71,10 +71,10 @@ class Connector implements ConnectorInterface
         // The following hack looks like the only way to
         // detect connection refused errors with PHP's stream sockets.
         if (false === stream_socket_get_name($socket, true)) {
-            return When::reject(new ConnectionException('Connection refused'));
+            return Promise\reject(new ConnectionException('Connection refused'));
         }
 
-        return When::resolve($socket);
+        return Promise\resolve($socket);
     }
 
     public function handleConnectedSocket($socket)
@@ -94,7 +94,7 @@ class Connector implements ConnectorInterface
     protected function resolveHostname($host)
     {
         if (false !== filter_var($host, FILTER_VALIDATE_IP)) {
-            return When::resolve($host);
+            return Promise\resolve($host);
         }
 
         return $this->resolver->resolve($host);